--- a/src/_BSplineBasis.jl
+++ b/src/_BSplineBasis.jl
@@ -257,7 +257,7 @@ end
     Expr(:block,
         :($(Expr(:meta, :inline))),
         :(k = knotvector(P)),
-        :($b = bsplinebasisall(_lower(P),i+1,t)),
+        :($b = bsplinebasisall(_lower_R(P),i+1,t)),
         :($(Bs[1]) = $(K1s[1])*$(bs[1])),
         exs...,
         :($(Bs[p+1]) = $(K2s[p])*$(bs[p])),