This PR fixes https://github.com/hyrodium/BasicBSpline.jl/issues/177 finally!