@hyrodium 2023-09-17
Finally! :tada:
```julia
julia> using BasicBSpline

julia> p = 3
3

julia> P1 = BSplineSpace{p}(KnotVector(1:8))
BSplineSpace{3, Int64, KnotVector{Int64}}(KnotVector([1, 2, 3, 4, 5, 6, 7, 8]))

julia> P2 = BSplineSpace{p}(KnotVector([2,3,3,4,5,6,7,8]))
BSplineSpace{3, Int64, KnotVector{Int64}}(KnotVector([2, 3, 3, 4, 5, 6, 7, 8]))

julia> changebasis(P1, P2)
4×4 SparseArrays.SparseMatrixCSC{Float64, Int32} with 5 stored entries:
 0.666667   ⋅    ⋅    ⋅ 
 0.333333  1.0   ⋅    ⋅ 
  ⋅         ⋅   1.0   ⋅ 
  ⋅         ⋅    ⋅   1.0

julia> P = BSplineSpace{p}(KnotVector(1:10))
BSplineSpace{3, Int64, KnotVector{Int64}}(KnotVector([1, 2, 3, 4, 5, 6, 7, 8, 9, 10]))

julia> P′ = BSplineSpace{p}(KnotVector([1,2,3,4,4.5,5,6,7,8.1,9.1,10.5]))
BSplineSpace{3, Float64, KnotVector{Float64}}(KnotVector([1.0, 2.0, 3.0, 4.0, 4.5, 5.0, 6.0, 7.0, 8.1, 9.1, 10.5]))

julia> changebasis(P,P′)
6×7 SparseArrays.SparseMatrixCSC{Float64, Int32} with 13 stored entries:
 1.0  0.166667   ⋅    ⋅         ⋅     ⋅           ⋅ 
  ⋅   0.833333  0.5   ⋅         ⋅     ⋅           ⋅ 
  ⋅    ⋅        0.5  0.833333   ⋅     ⋅           ⋅ 
  ⋅    ⋅         ⋅   0.166667  1.0  -0.0333333   0.0183333
  ⋅    ⋅         ⋅    ⋅         ⋅    1.03333    -0.103333
  ⋅    ⋅         ⋅    ⋅         ⋅     ⋅          1.085
```