@hyrodium 2023-09-17
Performance improvements on `changebasis_R`:tada:

**Before this PR**
```julia
julia> using BasicBSpline, BenchmarkTools

julia> k1 = knotvector"4 1 2 32 1111"
KnotVector([1, 1, 1, 1, 3, 5, 5, 7, 7, 7, 8, 8, 10, 11, 12, 13])

julia> k2 = knotvector"4321214311111"
KnotVector([1, 1, 1, 1, 2, 2, 2, 3, 3, 4, 5, 5, 6, 7, 7, 7, 7, 8, 8, 8, 9, 10, 11, 12, 13])

julia> P1 = BSplineSpace{3}(k1)
BSplineSpace{3, Int64, KnotVector{Int64}}(KnotVector([1, 1, 1, 1, 3, 5, 5, 7, 7, 7, 8, 8, 10, 11, 12, 13]))

julia> P2 = BSplineSpace{3}(k2)
BSplineSpace{3, Int64, KnotVector{Int64}}(KnotVector([1, 1, 1, 1, 2, 2, 2, 3, 3, 4, 5, 5, 6, 7, 7, 7, 7, 8, 8, 8, 9, 10, 11, 12, 13]))

julia> P1 ⊆ P2
true

julia> changebasis(P1, P2)
12×21 SparseArrays.SparseMatrixCSC{Float64, Int32} with 42 stored entries:
⎡⠙⢿⣶⣦⣀⠀⠀⠀⠀⠀⠀⎤
⎢⠀⠀⠀⠀⠉⠳⠤⡀⠀⠀⠀⎥
⎣⠀⠀⠀⠀⠀⠀⠀⠈⠛⢦⡀⎦

julia> @benchmark changebasis(P1, P2)
BenchmarkTools.Trial: 10000 samples with 3 evaluations.
 Range (min … max):   8.777 μs … 712.151 μs  ┊ GC (min … max): 0.00% … 95.12%
 Time  (median):      9.545 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   10.426 μs ±  16.585 μs  ┊ GC (mean ± σ):  4.05% ±  2.51%

    ▃▅▇███▇▆▅▄▃▁    ▁ ▁▁▁▁▁▂▂▃▂▃▂▂▂▂▁▁▁                        ▂
  ▆███████████████▇███████████████████████▇█▇▇▇▆▆▇▆▆▆▇▇▆▇▆▇▆▆▅ █
  8.78 μs       Histogram: log(frequency) by time      14.9 μs <

 Memory estimate: 11.09 KiB, allocs estimate: 76.
```

**After this PR**
```julia
julia> using BasicBSpline, BenchmarkTools

julia> k1 = knotvector"4 1 2 32 1111"
KnotVector([1, 1, 1, 1, 3, 5, 5, 7, 7, 7, 8, 8, 10, 11, 12, 13])

julia> k2 = knotvector"4321214311111"
KnotVector([1, 1, 1, 1, 2, 2, 2, 3, 3, 4, 5, 5, 6, 7, 7, 7, 7, 8, 8, 8, 9, 10, 11, 12, 13])

julia> P1 = BSplineSpace{3}(k1)
BSplineSpace{3, Int64, KnotVector{Int64}}(KnotVector([1, 1, 1, 1, 3, 5, 5, 7, 7, 7, 8, 8, 10, 11, 12, 13]))

julia> P2 = BSplineSpace{3}(k2)
BSplineSpace{3, Int64, KnotVector{Int64}}(KnotVector([1, 1, 1, 1, 2, 2, 2, 3, 3, 4, 5, 5, 6, 7, 7, 7, 7, 8, 8, 8, 9, 10, 11, 12, 13]))

julia> P1 ⊆ P2
true

julia> changebasis(P1, P2)
12×21 SparseArrays.SparseMatrixCSC{Float64, Int32} with 42 stored entries:
⎡⠙⢿⣶⣦⣀⠀⠀⠀⠀⠀⠀⎤
⎢⠀⠀⠀⠀⠉⠳⠤⡀⠀⠀⠀⎥
⎣⠀⠀⠀⠀⠀⠀⠀⠈⠛⢦⡀⎦

julia> @benchmark changebasis(P1, P2)
BenchmarkTools.Trial: 10000 samples with 6 evaluations.
 Range (min … max):  5.891 μs … 355.995 μs  ┊ GC (min … max): 0.00% … 94.82%
 Time  (median):     6.509 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   7.387 μs ±  11.802 μs  ┊ GC (mean ± σ):  5.75% ±  3.54%

   ▁▄▇███▇▆▄▃▂▁▂▃▄▄▄▃▃▂▁    ▁  ▁▁▁▁ ▁     ▁                   ▂
  ▇█████████████████████████████████████████████▇▇▇▇▇▆▆▅▆▆▆▆▅ █
  5.89 μs      Histogram: log(frequency) by time      11.4 μs <

 Memory estimate: 11.00 KiB, allocs estimate: 70.
```