@hyrodium 2023-09-18
The performance of `chagangebasis_I` is getting worse. This will be fixed in the future PRs.

**Before this PR**
```julia
julia> using BasicBSpline, BenchmarkTools

julia> p = 3
3

julia> P = BSplineSpace{p}(KnotVector(1:10))
BSplineSpace{3, Int64, KnotVector{Int64}}(KnotVector([1, 2, 3, 4, 5, 6, 7, 8, 9, 10]))

julia> P′ = BSplineSpace{p}(KnotVector([1,2,3,4,4.5,5,6,7,8.1,9.1,10.5]))
BSplineSpace{3, Float64, KnotVector{Float64}}(KnotVector([1.0, 2.0, 3.0, 4.0, 4.5, 5.0, 6.0, 7.0, 8.1, 9.1, 10.5]))

julia> changebasis(P,P′)
6×7 SparseArrays.SparseMatrixCSC{Float64, Int32} with 16 stored entries:
 1.0          0.166667   ⋅    ⋅         ⋅     ⋅           ⋅ 
 8.88178e-16  0.833333  0.5   ⋅         ⋅     ⋅           ⋅ 
 4.44089e-16  0.0       0.5  0.833333   ⋅     ⋅           ⋅ 
  ⋅            ⋅         ⋅   0.166667  1.0  -0.0333333   0.0183333
  ⋅            ⋅         ⋅    ⋅         ⋅    1.03333    -0.103333
  ⋅            ⋅         ⋅    ⋅         ⋅     ⋅          1.085

julia> @benchmark changebasis(P,P′)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):   9.387 μs …  3.547 ms  ┊ GC (min … max): 0.00% … 97.65%
 Time  (median):     10.710 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   12.354 μs ± 59.941 μs  ┊ GC (mean ± σ):  8.23% ±  1.69%

     ▁▅▇█▅▅▂▁                                                  
  ▁▂▅████████▇▅▄▃▃▃▃▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  9.39 μs         Histogram: frequency by time        18.7 μs <

 Memory estimate: 13.78 KiB, allocs estimate: 153.
```

**After this PR**
```julia
julia> using BasicBSpline, BenchmarkTools

julia> p = 3
3

julia> P = BSplineSpace{p}(KnotVector(1:10))
BSplineSpace{3, Int64, KnotVector{Int64}}(KnotVector([1, 2, 3, 4, 5, 6, 7, 8, 9, 10]))

julia> P′ = BSplineSpace{p}(KnotVector([1,2,3,4,4.5,5,6,7,8.1,9.1,10.5]))
BSplineSpace{3, Float64, KnotVector{Float64}}(KnotVector([1.0, 2.0, 3.0, 4.0, 4.5, 5.0, 6.0, 7.0, 8.1, 9.1, 10.5]))

julia> changebasis(P,P′)
6×7 SparseArrays.SparseMatrixCSC{Float64, Int32} with 13 stored entries:
 1.0  0.166667   ⋅    ⋅         ⋅     ⋅           ⋅ 
  ⋅   0.833333  0.5   ⋅         ⋅     ⋅           ⋅ 
  ⋅    ⋅        0.5  0.833333   ⋅     ⋅           ⋅ 
  ⋅    ⋅         ⋅   0.166667  1.0  -0.0333333   0.0183333
  ⋅    ⋅         ⋅    ⋅         ⋅    1.03333    -0.103333
  ⋅    ⋅         ⋅    ⋅         ⋅     ⋅          1.085

julia> @benchmark changebasis(P,P′)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):   87.757 μs …   6.008 ms  ┊ GC (min … max):  0.00% … 94.09%
 Time  (median):      93.898 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   109.354 μs ± 225.588 μs  ┊ GC (mean ± σ):  11.39% ±  5.40%

   ▁▅▇█▇▆▅▃▃▂▂▁▁▂▂▃▃▂▁▁▁▁                                       ▂
  ▅█████████████████████████▇▇▇▆▆▄▆▅▃▄▅▄▁▃▃▄▄▃▅▄▁▄▃▄▃▃▃▄▁▁▄▁▁▁▅ █
  87.8 μs       Histogram: log(frequency) by time        162 μs <

 Memory estimate: 145.64 KiB, allocs estimate: 1663.
```